ectcs :: trying to exclude .jpg, but not working

+++ src/ectsc.py
@@ -20,7 +20,7 @@ import ectcommon as co
 max_chunks = 1  # The sample-size defining short-hash, where long-hash always is the entire file.
 
 
-def make_rootlists(db):
+def make_rootlists_OLD(db):
     """ Look at argv and make a 1 element root list of it, if missing make root list from db """
     lst_i, lst_x = list(), list()
     if len(sys.argv) > 1:
@@ -36,7 +36,29 @@ def make_rootlists(db):
                 print(f"Unexpected mode: {row[0]} in {row}")
     return lst_i, lst_x
 
+
+def make_rootlists(db):
+    """ Look at argv and make a 1 element root list of it, if missing make root list from db """
+    lst_i, lst_x, lst_xe = list(), list(), list()
+    if len(sys.argv) > 1:
+        if os.path.isdir(sys.argv[1]):
+            lst_i.append(sys.argv[1])
+    else:  # build from db
+        for row in db.execute('SELECT * FROM roots'):
+            if row[0].upper() == 'I':  # include
+                lst_i.append(row[1])
+            elif row[0].upper() == 'X':  # exclude
+                lst_x.append(row[1])
+            elif row[0].upper() == 'XE':  # exclude extension
+                lst_xe.append(row[1])
+            else:
+                print(f"Unexpected mode: {row[0]} in {row}")
+    return lst_i, lst_x, lst_xe
+
+
 def timeandsize(str_ffn):
+    """ Return the file-time and file-size for file str_ffn, if exist
+        else return "", "" """
     try:
         str_filetime = time.strftime('%Y-%m-%dT%H:%M:%S', time.localtime(os.path.getmtime(str_ffn)))
         str_filesize = os.path.getsize(str_ffn)
@@ -44,7 +66,10 @@ def timeandsize(str_ffn):
     except FileNotFoundError:  # some files are very temporary ...
         return "", ""
 
+
 def scan_file(str_ffn):
+    # if "Glasvinge" in str_ffn:
+    #     print(f"G: {str_ffn}")
     str_scantime = datetime.datetime.now().isoformat()
     str_filetime, str_filesize = timeandsize(str_ffn)
     try:
@@ -53,6 +78,7 @@ def scan_file(str_ffn):
         return str_filetime, str_filesize, str_shorthas, str_longhash, str_scantime
     except FileNotFoundError:  # some files are very temporary ...
         return "", "", "", "", str_scantime
+
 
 def add_file2db(str_ffn, db):
     # print(f"Add file: {str_ffn}")
@@ -67,6 +93,7 @@ def add_file2db(str_ffn, db):
     except FileNotFoundError:
         pass  # some files are very temporary ...
 
+
 def add_longhash_2_shorthash(shorthash, db):
     str_sql_sel = f"SELECT * FROM files WHERE shorthash = '{shorthash}'"
     for row in db.execute(str_sql_sel):
@@ -78,14 +105,15 @@ def add_longhash_2_shorthash(shorthash, db):
             db.execute(str_sql_upd)
             db.commit()
 
-def scan_root(str_ri, lst_rx, db):
+
+def scan_root(str_ri, lst_rx, lst_rxe, db):
     """
     Scan one root-dir lst_ri, but skipping sub-dirs listed in lst_rx
     :param str_ri: string, the root dir
     :param lst_rx: list of exceptions, sub dirs that should be skipped
     :return: tbd
     """
-    print(f"scan_root(); Scanning: {str_ri, lst_rx}")
+    print(f"scan_root(); Scanning: {str_ri, lst_rx, lst_rxe}")
     # get db info on this dir
     dic_db = dict()  # dic by ffn of files known to the db
     str_sql = f"SELECT * FROM files where filename like '{str_ri}%'"
@@ -106,27 +134,31 @@ def scan_root(str_ri, lst_rx, db):
     num_cntfil = 0
     for root, dirs, files in os.walk(str_ri):
         for str_fn in files:
+            # if str_fn.lower().endswith('.jpg'):
+            #     print(str_fn)
             num_cntfil += 1
-            str_ffn = os.path.join(root, str_fn)
-            if not any([str_ffn.startswith(x) for x in lst_rx]):  # if the file is not excluded
-                if str_ffn in dic_db.keys():  # db knows this file
-                    obj_bdg = dic_db[str_ffn]
-                    tim, siz = timeandsize(str_ffn)
-                    if tim == dic_db[str_ffn][1] and siz == dic_db[str_ffn][2]:
-                        pass  # print(f" - skipping known file: {str_ffn} == {dic_db[str_ffn]}")  #
-                    else:
-                        ## print(f"WTF: tim? {tim == dic_db[str_ffn][1]} siz? {siz == dic_db[str_ffn][2]} @ ffn: {str_ffn}")
-                        # time or date have changed - so re-scanning file, and update DB.
-                        str_sql = f"DELETE FROM files WHERE filename='{str_ffn}';"
-                        db.execute(str_sql)
-                        db.commit()
+            if not any([str_fn.endswith(e) for e in lst_rxe]):
+                str_ffn = os.path.join(root, str_fn)
+                if not any([str_ffn.startswith(x) for x in lst_rx]):  # if the file is not excluded
+                    if str_ffn in dic_db.keys():  # db knows this file
+                        obj_bdg = dic_db[str_ffn]
+                        tim, siz = timeandsize(str_ffn)
+                        if tim == dic_db[str_ffn][1] and siz == dic_db[str_ffn][2]:
+                            pass  # print(f" - skipping known file: {str_ffn} == {dic_db[str_ffn]}")  #
+                        else:
+                            ## print(f"WTF: tim? {tim == dic_db[str_ffn][1]} siz? {siz == dic_db[str_ffn][2]} @ ffn: {str_ffn}")
+                            # time or date have changed - so re-scanning file, and update DB.
+                            str_sql = f"DELETE FROM files WHERE filename='{str_ffn}';"
+                            db.execute(str_sql)
+                            db.commit()
+                            add_file2db(str_ffn, db)
+                    else:  # db don't know this file - add it.
                         add_file2db(str_ffn, db)
-                else:  # db don't know this file - add it.
-                    add_file2db(str_ffn, db)
-            if num_cntfil % 1000000 == 0:
-                print(f"Count: {num_cntfil}: {str_ffn}")
+                if num_cntfil % 1000000 == 0:
+                    print(f"Count: {num_cntfil}: {str_ffn}")
 
-def scan_rootlists(lst_ri, lst_rx, db):
+
+def scan_rootlists(lst_ri, lst_rx, lst_rxe, db):
     """
     Scan all roots in rootlist lst_ri, though never the exceptions in lst_rx
     :param lst_ri:
@@ -136,7 +168,7 @@ def scan_rootlists(lst_ri, lst_rx, db):
     for root in lst_ri:
         lst_rxs = [tok for tok in lst_rx if tok.find(root) >= 0]
         print(f"\nscan_rootlists(); Clear to scan: ri: {root}, xs: {lst_rxs}")
-        scan_root(root, lst_rxs, db)
+        scan_root(root, lst_rxs, lst_rxe, db)
 
 
 def prioritize_candidates(lst_cand):
@@ -157,7 +189,7 @@ def prioritize_candidates(lst_cand):
                 cand[0] += 10
             if cand[1].lower().find("serie") > -1:
                 cand[0] += 10
-            if cand[1].find("__NAM__") > -1:
+            if cand[1].find("__NAM") > -1:
                 cand[0] += 10
             if cand[1].find("BIX_") > -1:
                 cand[0] += 10
@@ -191,6 +223,7 @@ def prioritize_candidates(lst_cand):
     else:  # Too few to prioritize
         return lst_cand  # return unchanged
 
+
 def main():
 
     db = co.db()
@@ -198,9 +231,10 @@ def main():
     print(f"argvar: {sys.argv[1:]}")  # ToDo: Take argvar out of main()
 
     # fill/update the db
-    lst_ri, lst_rx = make_rootlists(db)
-    print(f"main(); Go w: {lst_ri, lst_rx}")
-    scan_rootlists(lst_ri, lst_rx, db)
+    lst_ri, lst_rx, lst_rxe = make_rootlists(db)
+    # lst_ri.append(r"/home/martin/MEGA/Private/Publications_My")
+    print(f"main(); Go w: {lst_ri, lst_rx, lst_rxe}")
+    scan_rootlists(lst_ri, lst_rx, lst_rxe, db)
 
     # Look for short collisions
     print(f"Look for short collisions...")
@@ -229,6 +263,7 @@ def main():
     print(f"space: {num_space}")
     for kill in lst_kill:
         print(f'rm "{kill}"')
+
 
 if __name__ == '__main__':
     main()
